[update]: cek id waktu delete

+++ pages/users/hapus.php
@@ -9,12 +9,22 @@ $id = $_GET['id'];
 
 // Delete user row from table based on given id
 if (isset($db)) {
-    $result = mysqli_query($db, "DELETE FROM users WHERE id='$id'");
-    if ($result) {
+    $check = mysqli_fetch_array(mysqli_query($db, "SELECT * FROM users WHERE id = '$id'"));
+    if ($check) {
+        $result = mysqli_query($db, "DELETE FROM users WHERE id='$id'");
         $_SESSION['toastr'] = array(
             'type' => 'success', // or 'success' or 'info' or 'warning'
             'message' => 'Data berhasil dihapus!',
             'title' => 'Sukses'
+        );
+        echo "<script>
+         window.location.href='../../index.php?page=users';
+         </script>";
+    } else {
+        $_SESSION['toastr'] = array(
+            'type' => 'warning', // or 'success' or 'info' or 'warning'
+            'message' => 'Data tidak ditemukan!',
+            'title' => 'Peringatan'
         );
         echo "<script>
          window.location.href='../../index.php?page=users';
